+++ src/main/java/com/school/StudyShare/community/entity/CommunityLike.java
@@ -14,11 +14,13 @@ public class CommunityLike {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "community_id")
-    private Community community; // 게시글 엔티티
+    private Community community;
 
-    private Long userId;
+    // 💡 [수정] Long -> Integer
+    private Integer userId;
 
-    public CommunityLike(Community community, Long userId) {
+    // 💡 [수정] 생성자도 Integer로 변경
+    public CommunityLike(Community community, Integer userId) {
         this.community = community;
         this.userId = userId;
     }
